# Optimize timeline features to minimize time and sql load