--- a/core/classes/IssueAssignedTimelineEvent.class.php
+++ b/core/classes/IssueAssignedTimelineEvent.class.php
@@ -38,7 +38,7 @@ class IssueAssignedTimelineEvent extends TimelineEvent {
 	 * @param integer $p_handler_id An user identifier.
 	 */
 	public function __construct( $p_timestamp, $p_user_id, $p_issue_id, $p_handler_id ) {
-		parent::__construct( $p_timestamp, $p_user_id, $p_issue_id );
+		parent::__construct( $p_timestamp, $p_user_id );
 
 		$this->issue_id = $p_issue_id;
 		$this->handler_id = $p_handler_id;