--- a/core/classes/IssueNoteCreatedTimelineEvent.class.php
+++ b/core/classes/IssueNoteCreatedTimelineEvent.class.php
@@ -38,7 +38,7 @@ class IssueNoteCreatedTimelineEvent extends TimelineEvent {
 	 * @param integer $p_issue_note_id A issue note identifier.
 	 */
 	public function __construct( $p_timestamp, $p_user_id, $p_issue_id, $p_issue_note_id ) {
-		parent::__construct( $p_timestamp, $p_user_id, $p_issue_id );
+		parent::__construct( $p_timestamp, $p_user_id );
 
 		$this->issue_id = $p_issue_id;
 		$this->issue_note_id = $p_issue_note_id;