This change uses now works directly on bug history table rows and only processes the number of rows
necessary to render the number of rows required in the UI.  This significantly reduces the load in the
default view which doesn't show all rows.  The big savings happen in busy weeks, with about parity for
normal weeks in terms of load.

**Date Range based on mantisbt.org dataset**
2014-12-05 .. 2014-12-12 [Prev] [Next]

**After Optimization (collapsed)**
Total queries executed: 54
Unique queries executed: 54
Total query execution time: 0.0746 seconds

**Before Optimization (collapsed)**
Total queries executed: 515
Unique queries executed: 515
Total query execution time: 0.2072 seconds

**After Optimization (all)**
Total queries executed: 514
Unique queries executed: 514
Total query execution time: 0.1747 seconds

**Before Optimization (all)**
Total queries executed: 515
Unique queries executed: 515
Total query execution time: 0.2086 seconds

Fixes [#18015](https://www.mantisbt.org/bugs/view.php?id=18015)