@dregad 2015-07-06
Nice work. Review comments:

The change in history_get_raw_events_array() feels a bit like a hack, especially the bit about the returned array's structure changing depending on input parameters. Not a huge deal, but maybe it's a sign that we need a new, dedicated API function ?

I'm also concerned about what would happen for a low-privileged user in a "tight" environment where many history records could get filtered out... In such case, it may be worth having a loop to "fill" the timeline up to *MAX_EVENTS*, what do you think ?

And a few nitpicks:

- Commit d5bf4071b5d878b44a97fbcd668378837bdf14ca: *should*, not "shoudl"
- Commit 8c9114e9f5cf511f646e43fc6326272df74a7800 message: "to work directory on history" --> to work *directly* on history ? Or am I missing something ?
- Commit messages lines are too long... [Should be wrapped at 72 chars](https://www.mantisbt.org/wiki/doku.php/mantisbt:git_commit_messages)