@vboctor 2015-07-06
@dregad I've done further refactoring that includes the cleanup work and guarantees accuracy in terms of number of events in the timeline and when to show the "show more" link.

**After Optimization (default)**
Total queries executed: 151
Unique queries executed: 151
Total query execution time: 0.0988 seconds

**After optimization (all)**
Total queries executed: 565
Unique queries executed: 565
Total query execution time: 0.1901 seconds

I'm not sure of an easy way to go back and revise commits that are not on the top of the branch.  So unless you really care, I won't be revising text in these.

@atrol Your comments are also addressed.