@dregad 2015-07-07
@vboctor thanks for the feedback and corrections. 

So it appears that with these additional changes, we're 
- slightly less efficient than the original code in terms of number of queries while performance remains about the same (slightly less in fact)
- 3 times more queries than initial fix with ~25ms longer execution time but I guess that's the expected price to pay for accuracy.

I'll try to have a more detailed look and give this a test run later on.