@atrol 2015-07-07
Would be fine to consider not just query execution time but also total page execution time.