@vboctor 2015-07-13
@dregad assuming that my comparison is accurate, there may be slight increase in # of queries / execution time as you noted.  Did you get a chance to test?  I would like to get this in.  We can always add further improvements later.